% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{ce}
\alias{ce}
\title{ce}
\usage{
ce(y, x)
}
\arguments{
\item{y}{Outcome (vector)}

\item{x}{Predictor (vector)}
}
\value{
A numerical value
}
\description{
Conditional Entropy, where: H(Y|X) = H(X∩Y)-H(X), P's = probability, H's = entropy, J's = joint probabilities
}
\examples{
ce(misimdata$y,misimdata$x1)
}
