% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entfun.R
\name{entfun}
\alias{entfun}
\title{entfun}
\usage{
entfun(y, x)
}
\arguments{
\item{y}{Outcome (vector)}

\item{x}{Predictor (data.frame object)}
}
\value{
A data.frame object
}
\description{
Run N conditional entropy estimates against a single dependent variable
}
\examples{
entfun(misimdata$y,misimdata[,2:5])
}
