% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entfuns.R
\name{entfuns}
\alias{entfuns}
\title{entfuns}
\usage{
entfuns(y, x)
}
\arguments{
\item{y}{Outcome (vector)}

\item{x}{Predictor (data.frame object)}
}
\value{
A data.frame object
}
\description{
A diagnostic and descriptive tool that calculates probability, conditional probability, absolute difference between marginal and conditional probability, absolute difference between marginal and conditional entropy, proportional difference between marginal and conditional probability, and proportional difference between marginal and conditional entropy
}
\examples{
entfuns(misimdata$y,misimdata[,2:5])
}
