% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{entropy}
\usage{
entropy(x)
}
\arguments{
\item{x}{A binary vector (numeric coded as 0/1 or logical). Must be length >= 1.}
}
\value{
A numeric scalar giving the entropy of `x`.
}
\description{
Returns marginal entropy for binary variables
}
\examples{
entropy(misimdata$x1)
}
