% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{joint}
\alias{joint}
\title{joint}
\usage{
joint(y, x)
}
\arguments{
\item{y}{Outcome (vector)}

\item{x}{Predictor (vector)}
}
\value{
A numerical value
}
\description{
Joint probability, P(Y∩X)
}
\examples{
joint(misimdata$y,misimdata$x1)
}
