% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{mi}
\alias{mi}
\title{mi}
\usage{
mi(y, x)
}
\arguments{
\item{y}{Outcome vector (binary: 0/1 or logical).}

\item{x}{Predictor vector (binary: 0/1 or logical). Must be the same length
as `y`.}
}
\value{
A numeric scalar giving the mutual information between `x` and `y`
}
\description{
Computes the mutual information (MI) between an outcome `y` and
a predictor `x`, using the standard definition: \deqn{MI(X, Y) = H(X) + H(Y)
- H(X, Y),}
}
\examples{
mi(misimdata$y,misimdata$x1)
}
