% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi2.R
\name{mi2}
\alias{mi2}
\title{mi2}
\usage{
mi2(y, x)
}
\arguments{
\item{y}{Outcome (vector)}

\item{x}{Predictor (vector)}
}
\value{
A numerical value
}
\description{
Additional formulation, where: I(X, Y) = H(Y) − H(Y|X), P's = probability, H's = entropy, J's = joint probabilities
}
\examples{
mi2(misimdata$y,misimdata$x1)
}
