% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misimdata}
\alias{misimdata}
\title{misimdata}
\format{
A data frame with 2500 rows and 6 variables:
\describe{
  \item{y}{Outcome}
  \item{x1}{First binary predictor}
  \item{x1}{Second binary predictor}
  \item{x1}{Third binary predictor}
  \item{x1}{Fourth binary predictor}
  \item{x1}{Fifth binary predictor}
}
}
\source{
<https://www.github.com/mvuorre/exampleRPackage>
}
\usage{
misimdata
}
\description{
A data set with 5 predictors and 1 outcome that can be used to try out the functions of the mitools package
}
\keyword{datasets}
