% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prob.R
\name{plot_prob}
\alias{plot_prob}
\title{plot_prob}
\usage{
plot_prob(
  data,
  y_var,
  x_vars,
  var_labels = NULL,
  prob_digits = 2,
  method = "conditional",
  title = NULL,
  vertex_color = "lightblue",
  vertex_frame_color = "darkblue",
  vertex_label_color = "black",
  edge_color = "darkgrey",
  edge_label_color = "black",
  min_arrow_width = 1,
  max_arrow_width = 10,
  node_size = 40,
  label_cex = 0.8
)
}
\arguments{
\item{data}{Dataset containing the yvar and the xvar (a data.frame object)}

\item{y_var}{Outcome (character)}

\item{x_vars}{A vector of strings indicating the predictors for the outcome (vector)}

\item{var_labels}{Labels for the xvars in the graph (vector)}

\item{prob_digits}{Number of decimals to round the conditional probability with, defaults to 2 (Integer)}

\item{method}{String that indicates whether conditional probabilities or logistic regression effects should be calculated (character)}

\item{title}{Title of the plot (character)}

\item{vertex_color}{String setting the color of the vertices (character)}

\item{vertex_frame_color}{String setting the color of the frames of the vertices (character)}

\item{vertex_label_color}{String setting the color of the labels of the vertices (character)}

\item{edge_color}{String setting the color of the edges (character)}

\item{edge_label_color}{String setting the color of labels the edges (character)}

\item{min_arrow_width}{Number used to define the minimum width of the arrows in the graph (Integer)}

\item{max_arrow_width}{Number used to define the maximum width of the arrows in the graph (Integer)}

\item{node_size}{Number used to define the size of the nodes  in the graph (Integer)}

\item{label_cex}{Number used to define the size of the node labels in the grap (Integer)}
}
\value{
A graph where the xvars and yvar are the nodes and the edges represent the relation between the xvars and the yvar.
On each node, the name of the variable and the marginal probability of the variable is displayed
On each edge, the conditional probability or logistic regression effect between a certain xvar and the yvar is displayed
}
\description{
Create a graph showing how sets of variables are related to an outcome via conditional probability or logistic regression effect.
}
\examples{
plot_prob(misimdata,'y',colnames(misimdata[,3:6]),method='logistic')
}
