% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probstat.R
\name{probstat}
\alias{probstat}
\title{probstat}
\usage{
probstat(y, x, nfolds, seed = 10101)
}
\arguments{
\item{y}{Outcome (vector)}

\item{x}{Predictor (a data.frame object, typically the expanded dataset from the pairmi function)}

\item{nfolds}{Number of folds for the cross-validation (Integer)}

\item{seed}{Number used for the randomization across folds (Integer)}
}
\value{
A data.frame object giving the following parameters for each pair:
xprob = probability of X
yprob = probability of Y
cprob = conditional probability of Y given X
cprobx = = conditional probability of X given Y
cprobi = inverse conditional probability of Y given X
cpdif = difference between marginal and conditional probability of Y
cpdifper = percentagewise difference between marginal and conditional probability of Y
xent = entropy of X
yent = entropy of Y
ce = conditional entropy
cedif = difference between marginal and conditional entropy of Y
cedifper = percentagewise difference between marginal and conditional entropy of Y
}
\description{
probstat
}
\examples{
probstat(misimdata$y,pairmiresult$expanded.data,5)
}
