% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setmapmi.R
\name{setmapmi}
\alias{setmapmi}
\title{setmapmi}
\usage{
setmapmi(original.variables = NULL, sets = NULL, n_elements = NULL)
}
\arguments{
\item{original.variables}{The names of the original variables that were paired up (Vector of strings)}

\item{sets}{The information on the sets from the pairmi function (Data.frame object)}

\item{n_elements}{The depth of sets that you want to visualize (Integer)}
}
\value{
A setmap showing which original variables make up the sets at a certain depth
}
\description{
A function that creates a setmap based on the output of the pairmi function
}
\examples{
pairmiresult = pairmi(misimdata[,2:6])
setmapmi(pairmiresult$original.variables,pairmiresult$sets,2)
}
