% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprob.R
\name{cprob}
\alias{cprob}
\title{cprob}
\usage{
cprob(y, x)
}
\arguments{
\item{y}{Outcome (vector)}

\item{x}{Predictor (vector)}
}
\value{
A numerical value
}
\description{
Conditional Probability, where: P(Y|X) = P(Y∩X)/P(X=1)
}
\examples{
cprob(misimdata$y,misimdata$x1)
}
