% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprob_inv.R
\name{cprob_inv}
\alias{cprob_inv}
\title{cprob_inv}
\usage{
cprob_inv(y, x)
}
\arguments{
\item{y}{Outcome (vector)}

\item{x}{Predictor (vector)}
}
\value{
A numerical value
}
\description{
Conditional Probability of the inverse, where: P(Y|X) = P(Y∩X)/P(X=0)
}
\examples{
cprob_inv(misimdata$y,misimdata$x1)
}
