% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{entropy}
\usage{
entropy(x)
}
\arguments{
\item{x}{Variable (vector)}
}
\value{
A numerical value
}
\description{
Returns marginal entropy for binary variables
}
\examples{
entropy(misimdata$x1)
}
