% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_minimal_sets.R
\name{find_minimal_sets}
\alias{find_minimal_sets}
\title{find_minimal_sets}
\usage{
find_minimal_sets(str_vec, sep = "_")
}
\arguments{
\item{str_vec}{Vector of sets for which you want to find the minimally sufficient sets (vector)}

\item{sep}{The seperator that was used in the creation of the sets (character)}
}
\value{
A vector with the minimally sufficient sets
}
\description{
A function that finds the irreducible and nonexchangeable sets that account for all other sufficient sets at a certain conditional probability threshold
}
\examples{
pairmiresult = pairmi(misimdata[,2:6])
results_probstat <- probstat(misimdata$y,pairmiresult$expanded.data,nfolds=5)
find_minimal_sets(results_probstat$xvars[results_probstat$cprob >= 0.20])
}
