% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstat.R
\name{gstat}
\alias{gstat}
\title{gstat}
\usage{
gstat(mi, count)
}
\arguments{
\item{mi}{Mutual information (numerical value)}

\item{count}{Joint count (integer value)}
}
\value{
A numerical value
}
\description{
G-statistic
}
\examples{
gstat(mi(misimdata$y,misimdata$x1),jtct(misimdata$y,misimdata$x1))
}
