% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtct.R
\name{jtct}
\alias{jtct}
\title{jtct}
\usage{
jtct(y, x)
}
\arguments{
\item{y}{Outcome (vector)}

\item{x}{Predictor (vector)}
}
\value{
An integer value
}
\description{
Joint count function
}
\examples{
cprob_inv(misimdata$y,misimdata$x1)
}
