% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{mi}
\alias{mi}
\title{mi}
\usage{
mi(y, x)
}
\arguments{
\item{y}{Outcome (vector)}

\item{x}{Predictor (vector)}
}
\value{
A numerical value
}
\description{
Mutual Information, where: H(X)+H(Y)−H(X,Y), P's = probability, H's = entropy, J's = joint probabilities
}
\examples{
mi(misimdata$y,misimdata$x1)
}
