% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misimdata}
\alias{misimdata}
\title{misimdata}
\format{
A data frame with 2500 rows and 11 variables:
\describe{
  \item{y}{Outcome}
  \item{x1}{First binary predictor}
  \item{x2}{Second binary predictor}
  \item{x3}{Third binary predictor}
  \item{x4}{Fourth binary predictor}
  \item{x5}{Fifth binary predictor}
  \item{x6}{Sixth binary predictor}
  \item{x7}{Seventh binary predictor}
  \item{x8}{Eighth binary predictor}
  \item{x9}{Ninth binary predictor}
  \item{x10}{Tenth binary predictor}
}
}
\usage{
misimdata
}
\description{
A data set with 10 predictors and 1 outcome that can be used to try out the functions of the setweaver package
}
\keyword{datasets}
