% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairmi.R
\name{pairmi}
\alias{pairmi}
\title{pairmi}
\usage{
pairmi(data, p.threshold = 0.05, MI.threshold = NULL, depth = 5, sep = "_")
}
\arguments{
\item{data}{A data.frame object with the variables that you want to pair up}

\item{p.threshold}{A numerical value indicating the p-value that is used to determine the significance of a pair}

\item{MI.threshold}{A numerical value indicating the MI value that is used to determine the significance of a pair. Overrides the p-value}

\item{depth}{An integer indicating the maximum number of variables a pair can have}

\item{sep}{A string which will be used to seperate the variables in the pair name}
}
\value{
A list that includes
(1) A data.frame object with the original data and the data of the significant pairs
(2) A vector with the orignal variable names
(3) A data.frame object with information on the significant pairs
}
\description{
A function that calculates the mutual information for pairs of variables, calculates the G statistic, determines the significance of the pairs, and only keeps those that are significant.
}
\examples{
pairmi(misimdata[,2:6])
}
