% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairmi.R
\name{pairmi}
\alias{pairmi}
\title{pairmi}
\usage{
pairmi(data, alpha = 0.05, MI.threshold = NULL, n_elements = 5, sep = "_")
}
\arguments{
\item{data}{A data.frame object with the variables that you want to pair up}

\item{alpha}{A numerical value indicating the p-value that is used to
determine the significance of a set}

\item{MI.threshold}{A numerical value indicating the MI value that is used to
determine the significance of a set Overrides the p-value}

\item{n_elements}{An integer indicating the maximum number of variables a set
can have}

\item{sep}{A string which will be used to separate the variables in the pair
name}
}
\value{
A list that includes (1) A data.frame object with the original data
  and the data of the significant sets (2) A vector with the original
  variable names (3) A data.frame object with information on the significant
  sets
}
\description{
A function that calculates the mutual information for sets of
  variables, calculates the G statistic, determines the significance of the
  sets, and only keeps those that are significant.
}
\examples{
pairmi(misimdata[,2:6])
}
