% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setmapmi.R
\name{setmapmi}
\alias{setmapmi}
\title{setmapmi}
\usage{
setmapmi(orig.variables = NULL, pairs = NULL, depth = NULL)
}
\arguments{
\item{orig.variables}{The names of the original variables that were paired up (Vector of strings)}

\item{pairs}{The information on the pairs from the pairmi function (Data.frame object)}

\item{depth}{The depth of sets that you want to visualize (Integer)}
}
\value{
A setmap showing which original variables make up the sets at a certain depth
}
\description{
A function that creates a setmap based on the output of the pairmi function
}
\examples{
setmapmi(pairmiresults$orig.variables,pairmiresults$pairs,2)
}
