% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zprob.R
\name{zprob}
\alias{zprob}
\title{zprob}
\usage{
zprob(x)
}
\arguments{
\item{x}{Variable (vector)}
}
\value{
A numerical value
}
\description{
z score for proportion (probability) differing from zero. To insure that distress is occurring a nonzero proportion of time
}
\examples{
zprob(misimdata$x1)
}
